Develop

+++ src/app/admin-interests/admin-interests.component.ts
@@ -13,7 +13,9 @@ export class AdminInterestsComponent {
    btnTxt: string = "Agregar";
    interests: Interests[] = [];
    myInterests: Interests = new Interests();
-    constructor(public interestsService: InterestsService)
+   selectedInterestId: string | null = null;
+   
+   constructor(public interestsService: InterestsService)
     {
         console.log(this.interestsService);
         this.interestsService.getInterests().snapshotChanges().pipe(
@@ -29,21 +31,49 @@ export class AdminInterestsComponent {
     }
 
   AgregarInterests(){
-    console.log(this.interestsService);
+   if (this.selectedInterestId) {
+     
+      this.interestsService.updateInterests(this.selectedInterestId, this.myInterests).then(() => {
+        this.resetForm();
+        console.log('Updated item successfully!');
+      });
+    } else {
+      
+      this.interestsService.createInterests(this.myInterests).then(() => {
+        this.resetForm();
+        console.log('Created new item successfully!');
+      });
+    }
+      
+      /* console.log(this.interestsService);
     this.interestsService.createInterests(this.myInterests).then(() => {
        console.log('Created new item successfully!');
-    });
+    });*/
   }
 
   deleteInterests(id? :string){
-    this.interestsService.deleteInterests(id).then(() => {
-       console.log('Delete item successfully!');
+    if (window.confirm('¿Estás segura de que deseas eliminar este interés?')) {
+      this.interestsService.deleteInterests(id).then(() => {
+      console.log('Delete item successfully!');
     });
     console.log(id);
+    } else {
+      console.log('Eliminación cancelada.');
+    }
+      /*    this.interestsService.deleteInterests(id).then(() => {
+       console.log('Delete item successfully!');
+    });
+    console.log(id);*/
   }
 
-  updateInterests(id? :string){
-    alert('updating...');
+  updateInterests(interest :any){
+    this.myInterests = { descripcion: interest.descripcion };  
+    this.selectedInterestId = interest.id;
+    this.btnTxt = 'Update';
   }
-
+  resetForm() {
+    this.myInterests = new Interests();
+    this.selectedInterestId = null;
+    this.btnTxt = 'Agregar';
+  }
 }
